feat(bar): page ingrédients alignée sur la cuisine (composant partagé)
Refactore /app/ingredients/page.js (cuisine, 821 lignes) en composant
partagé components/IngredientsView.jsx paramétré par section. Les
routes /ingredients et /bar/ingredients deviennent des wrappers de 5
lignes qui passent section="cuisine" ou section="bar".

Ce que la page bar gagne, en parité avec la cuisine :
- KPIs en tête : Total / Catégories / Sans catégorie.
- Onglets Liste / Catégories / Inflation.
- Vue Liste : colonne Catégorie avec badge, colonne Action avec bouton
  Modifier (édition inline nom/prix/unité/catégorie), badge "À catégoriser"
  pour les ingrédients orphelins, pagination par 30.
- Vue Catégories : CRUD complet (création / édition emoji+nom / suppression
  avec détachement des ingrédients) — comble le manque "j'ai pas l'onglet
  pour créer des catégories".
- Vue Inflation : stats prix moyen/min/max/écart par catégorie.
- Formulaire d'ajout d'ingrédient avec sélecteur de catégorie.
- Couleurs passées au thème (branding tenant) au lieu du violet hardcodé,
  pour respecter la charte de l'établissement.

Les différences cuisine ↔ bar sont concentrées dans une const CONFIG :
- table : ingredients vs ingredients_bar
- relation fiches : fiche_ingredients vs fiche_bar_ingredients
- unité par défaut : kg vs cl
- rôles autorisés et redirection si mauvais rôle
- chemin d'import : /import vs /bar/import
- tables à détacher en cascade
